@@ -17,7 +17,7 @@
     </PropertyGroup>
 
     <PropertyGroup>
-        <LangVersion>9.0</LangVersion>
+        <LangVersion>11.0</LangVersion>
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
